feat: Creating new macros module and moving portal, and room macro there.
