+++ src/lib.rs
@@ -1,6 +1,8 @@
 pub mod ret_lang;
 pub mod game;
 pub mod migration;
+#[macro_use]
+pub mod macros;
 
 const DB_PATH: &str = "~/retribution.db";
